+++ rollup.config-1740292572290.cjs
@@ -1,0 +1,52 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var peerDepsExternal = require('rollup-plugin-peer-deps-external');
+var resolve = require('@rollup/plugin-node-resolve');
+var commonjs = require('@rollup/plugin-commonjs');
+var typescript = require('@rollup/plugin-typescript');
+var dts = require('rollup-plugin-dts');
+var terser = require('@rollup/plugin-terser');
+var postcss = require('rollup-plugin-postcss');
+
+const packageJson = require("./package.json");
+
+var rollup_config = [
+  {
+    input: "src/index.ts",
+    output: [
+      {
+        file: packageJson.main,
+        format: "cjs",
+        // sourceMap: true,
+      },
+      {
+        file: packageJson.module,
+        format: "esm",
+        // sourceMap: true,
+      },
+    ],
+    plugins: [
+      peerDepsExternal(),
+      resolve(),
+      commonjs(),
+      typescript({ tsconfig: "./tsconfig.json" }),
+      terser(),
+      postcss(),
+    ],
+    external: ["react", "react-dom"],
+  },
+  {
+    input: "src/index.ts",
+    output: [
+      {
+        file: packageJson.types,
+      },
+    ],
+    plugins: [dts.default()],
+    external: [/\.css/],
+  },
+];
+
+exports.default = rollup_config;