+++ src/index.ts
@@ -1,4 +1,5 @@
 import CalendarView from "./components/calendar";
 import DragToReOrder from "./components/drag-to-reorder";
 import ImageCarousel from "./components/carousel";
-export { CalendarView, DragToReOrder, ImageCarousel };
+import TriggerDownload from "./components/trigger-download";
+export { CalendarView, DragToReOrder, ImageCarousel, TriggerDownload };